web/requirements.txt
fastapi
uvicorn[standard]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
